Translate to Italian #5
The intent of this commit is to translate all displayed English text to
Italian.

It also uses a custom name for the app, and custom images & favicon.

Finally, a new component, Textual.tsx, is used to insert a note on the
right-hand side of the page about how to update meetings.

* public/index.html
Translate to Italian

* public/favicon.ico
* public/logo192.png
* public/logo512.png
Customise favicon, and app images

* public/manifest.json
Customise app name to Italian

* src/components/Filter.tsx
(Filter): import and use Textual component

* src/components/Meeting.tsx
(Meeting): Times in Italy are in the form 23:59 not 11:59 PM

* src/components/Search.tsx
(Search): translate to Italian

* src/components/Textual.tsx
(Textual): create a new component 'email here to update'

* src/helpers/config.ts
(dataUrl): use Italy meeting source spreadsheet
(days): translate days to Italian

* src/helpers/data.ts
Import Italy locale
(originalUrl): translate to Italian